server/requirements.txt
altgraph==0.17.2
astroid==2.8.4
attrs==21.4.0
autopep8==1.6.0
click==8.0.3
colorama==0.4.4
flasgger==0.9.5
Flask==2.0.2
future==0.18.2
isort==5.10.0
itsdangerous==2.0.1
Jinja2==3.0.2
jsonschema==4.4.0
lazy-object-proxy==1.6.0
MarkupSafe==2.0.1
mccabe==0.6.1
mistune==2.0.2
mypy==0.910
mypy-extensions==0.4.3
numpy==1.21.4
pandas==1.3.4
pefile==2021.9.3
platformdirs==2.4.0
pycodestyle==2.8.0
pyinstaller==4.6
pyinstaller-hooks-contrib==2021.3
PyJWT==2.3.0
pylint==2.11.1
pyrsistent==0.18.1
python-dateutil==2.8.2
pytz==2021.3
pywin32-ctypes==0.2.0
PyYAML==6.0
Rx==3.2.0
six==1.16.0
toml==0.10.2
typing==3.7.4.3
typing-extensions==3.10.0.2
Werkzeug==2.0.2
wrapt==1.13.3
